LayoutTests/inspector-obsolete/debugger/source-url-comment-expected.txt

Tests that evals with sourceURL comment are shown in scripts panel.

Debugger was enabled.

Running: testSourceURLCommentInInlineScript

Running: testSourceURLComment
function keepAlive() {}
//@ sourceURL=evalURL.js

Running: testSourceURLCommentInDynamicScript
function keepAliveInDynamicScript() {}
//@ sourceURL=dynamicScriptURL.js

Running: testNonRelativeURL
function relativeURLScript() {}
//@ sourceURL=/js/nonRelativeURL.js

Running: testHashSyntax
function hashSyntax() {}
//# sourceURL=hashSyntax.js
Debugger was disabled.
